scraper/requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
supabase==2.15.0
pytest==8.3.5
pytest-cov==6.1.1
flake8==7.1.1
bandit==1.8.3
certifi
